refactor(documents): zera react-doctor em DocumentUpload (Onda 4, #254)
Onda 4 — Hotspot da campanha "Zerar react-doctor" (épico #239). Separa
lógica (hook) de apresentação (componentes controlados), reduzindo o
componente de 404 para ~70 linhas de wiring.

Diagnósticos eliminados (react-doctor 0.5.6; 5 → 0 em DocumentUpload.tsx):
- no-giant-component     -> lógica movida p/ hook useDocumentUpload (sem JSX;
                           a regra é react-jsx-only, então o hook não é alvo)
- prefer-useReducer      -> estado reduzido a 3 useState (phase união
                           discriminada + csv + mapping); loading derivado de
                           phase. A união quebra o cluster step/progress/
                           analysis. Sem useReducer (vetado pela issue)
- async-await-in-loop #1 -> loop de checkDuplicates paralelizado com
                           Promise.all (chunks independentes, agregação
                           comutativa)
- async-await-in-loop #2 -> upload serial mantido de propósito (isLast
                           finaliza no backend + progresso sequencial),
                           suprimido inline e documentado em LINT_CONFIG.md
- no-array-index-as-key  -> key estável por linha em CsvPreviewTable

Corrige também o bug latente que a união discriminada expôs: uma falha de
upload num CSV sem duplicatas caía em setStep("analysis") com analysis=null,
renderizando painel em branco; agora restaura a fase de origem (mapping no
caminho direto, analysis quando veio do painel) — retry possível.

Novos: hooks/useDocumentUpload.ts, components/documents/{CsvDropzone,
MappingStep,CsvPreviewTable}.tsx. Comportamento de upload preservado.

Validação: 569 testes passam, tsc + eslint limpos, react-doctor sem
diagnóstico nos 5 arquivos.

Closes #254
